[REFACTOR]: indexPath.row 값 제대로 지정

+++ FenceApp/Modals/SettingModalViewController.swift
@@ -104,15 +104,17 @@ extension SettingModalViewController: UITableViewDelegate, UITableViewDataSource
                     print("실패")
                 }
             }
-        } else if indexPath.row == 2 {
+        } else if indexPath.row == 1 {
             let webViewController = WebViewController(urlString: "https://dandy-temple-d75.notion.site/884b394ea65d4cfd9047b2f1d5010839?pvs=4")
             self.present(webViewController, animated: true)
-        } else if indexPath.row == 3 {
+        } else if indexPath.row == 2 {
             let webViewController = WebViewController(urlString: "https://dandy-temple-d75.notion.site/eeffb1969f0b4c4383f0c5494e99b614?pvs=4")
             self.present(webViewController, animated: true)
         } else if indexPath.row == 3 {
             let alertController = UIAlertController(title: "로그아웃", message: "로그아웃 하시겠습니까?", preferredStyle: .alert)
+            
             let cancelAction = UIAlertAction(title: "취소", style: .cancel, handler: nil)
+            
             let confirmAction = UIAlertAction(title: "확인", style: .default) { [weak self] action in
                 
                 do {
@@ -140,6 +142,13 @@ extension SettingModalViewController: UITableViewDelegate, UITableViewDataSource
             let confirmAction = UIAlertAction(title: "확인", style: .default) { [weak self] action in
                 guard let email = alertController.textFields?[0].text else { return }
                 guard let password = alertController.textFields?[1].text else { return }
+                
+                guard email == CurrentUserInfo.shared.currentUser?.email else {
+                    let incorrectAlertController = UIAlertController(title: "이메일이 틀립니다", message: nil, preferredStyle: .alert)
+                    incorrectAlertController.addAction(cancelAction)
+                    self?.present(incorrectAlertController, animated: true)
+                    return
+                }
                 Task {
                     do {
                         try await self?.firebaseAuthService.deleteUser(email: email, password: password)
